Specification

IntervalMDPAbstractions.FiniteTimeRegionReachability — Type

FiniteTimeRegionReachability

A struct representing a finite-time reachability property.

IntervalMDPAbstractions.InfiniteTimeRegionReachability — Type

InfiniteTimeRegionReachability

A struct representing a infinite-time reachability property.

IntervalMDPAbstractions.FiniteTimeRegionReachAvoid — Type

FiniteTimeRegionReachAvoid

A struct representing a finite-time reach-avoid property.

IntervalMDPAbstractions.InfiniteTimeRegionReachAvoid — Type

InfiniteTimeRegionReachAvoid

A struct representing a infinite-time reach-avoid property.

IntervalMDPAbstractions.FiniteTimeRegionSafety — Type

FiniteTimeRegionSafety

A struct representing a finite-time safety property.

IntervalMDPAbstractions.InfiniteTimeRegionSafety — Type

InfiniteTimeRegionSafety

A struct representing a infinite-time safety property.

IntervalMDPAbstractions.reach — Function

reach

Return the reach region of a reachability or reach-avoid property.

IntervalMDPAbstractions.avoid — Function

avoid

Return the avoid region of a reach-avoid or safety property.

IntervalMDPAbstractions.dim — Function

dim

Return the dimension of the reach and avoid regions of a regional property.

Abstraction problem

IntervalMDPAbstractions.AbstractionProblem — Type

AbstractionProblem

A struct of a system and a specification to be used in the abstraction process.

IntervalMDPAbstractions.system — Function

system

Return the system of an abstraction problem.

IntervalMDPAbstractions.specification — Function

specification

Return the specification of an abstraction problem.